Added Custom Embed

+++ bot.py
@@ -8,25 +8,29 @@ load_dotenv() # load variables from .env file
 TOKEN = os.environ.get('DISCORD_TOKEN') # get the token from the environment variables
 
 
-# Embed Message
 async def send_message(message, user_message, is_private):
     try:
         if user_message.startswith('embed'):
-            embed = discord.Embed(title="Guten Tag!", description="This is a test message.", color=discord.Color.green())
-            await message.channel.send(embed=embed)
+            args = user_message.split('|')
+            title = args[0].split('=')[1].strip()
+            description = args[1].split('=')[1].strip()
+            embed = discord.Embed(title=title, description=description, color=0x00ff00)
+            sent_msg = await message.channel.send(embed=embed)
+            await message.delete()
         else:
             response = responses.get_response(user_message)
             if response: # only send a reply if the response is not empty
                 await message.author.send(response) if is_private else await message.channel.send(response)
+            await message.delete()
 
     except Exception as e:
         print(e)
 
-    
+
 
 async def on_offend(msg, client):
 
-    block_words = ["white", "black", "http://", "https://"]
+    block_words = ["black", "Black", "http://", "https://"]
     # If the message wasn't sent by the bot
     if msg.author != client.user:
 
@@ -34,12 +38,11 @@ async def on_offend(msg, client):
         for text in block_words:
             # Checking if the message was sent by a moderator (because it would be nice if moderators
             # could share links in case it would be important for them to do so).
-            if "Moderator" not in str(msg.author.roles) and text in str(msg.content.lower()):
+            if "Moderator" not in str(msg.author.roles) and "admin" not in [role.name.lower() for role in msg.author.roles] and text in str(msg.content.lower()):
                 await msg.delete() # Deletes the message
                 return # So that we don't continue going throuh the loop once we've already found
                        # a blocked word
 
-        print("Not Deleting...") # This will be printed if the bot doesn't delete a message
 
 # Bot Setup
 def run_discord_bot():
